refactor the codes by following SOLID, KISS, and DRY method and leave only code to run the app in the main file
